Improved tests folder structure
