+++ src/components/Modal/modal.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { render } from 'test-utils';
+
+import Modal from './Modal';
+
+test('Modal component renders', () => {
+  render(<Modal />);
+});
