+++ src/components/Verb/index.js
@@ -105,7 +105,6 @@ class Verb extends Component {
 
   render() {
     const { randomVerb, randomPerson, data, count } = this.state;
-    console.log('Answer:', randomPerson[1]);
     return (
       <div>
         <Input
